Sanitized Input for Topics
Sanitized and validated topic input ranging from id lengths, text length, and text content with Regex.

+++ Backend/modules/topic.js
@@ -10,6 +10,26 @@ const { gen_topic_id } = require('../lib/id_gen');
 
 // LOCAL FUNCTIONS
 
+/**
+ * Check topic title for being empty or greater than defined length.
+ * Validate input values.
+ * @param {String} topic_title 
+ */
+function check_title(topic_title) {
+    if (!topic_title || (topic_title.length > 64)) return false;
+    return /^[a-z0-9]+$/i.test(topic_title);
+}
+
+/**
+ * Check topic title for being empty, or having valid characters,
+ * additionally check that length isn't excessive (> 500).
+ * @param {string} topic_content 
+ */
+function check_content(topic_content) {
+    if (!topic_content || (topic_content.length > 500)) return false;
+    return /^[a-z0-9]+$/i.test(topic_content);
+}
+
 // EXPORTED FUNCTIONS
 
 /**
@@ -19,6 +39,18 @@ const { gen_topic_id } = require('../lib/id_gen');
  */
 exports.CreateTopic =
 async function CreateTopic(userID, topic_title, topic_type, topic_content) {
+    // Validate & Sanitize input.
+    if (!check_title(topic_title) || !check_content(topic_content)) {
+        return {
+            success: false,
+            server_error: false,
+            invalid_input: true,
+        };
+    }
+    // Sanitize topic title and input.
+    topic_title = topic_title.replace(/[^a-zA-Z0-9\w\s]/gi, "");
+    topic_content = topic_content.replace(/[^a-zA-Z0-9\w\s]/gi, "");
+    // Send DB query.
     let topic_id = gen_topic_id();  // Generate topic ID.
     let topic = {
         topic_id: topic_id,
@@ -42,7 +74,6 @@ async function CreateTopic(userID, topic_title, topic_type, topic_content) {
         success: true,
         topic: {
             topic_id: topic_id,
-            //time_posted: //TODO
         }
     };
 }
@@ -53,8 +84,9 @@ async function CreateTopic(userID, topic_title, topic_type, topic_content) {
  */
 exports.DeleteTopic =
 async function DeleteTopic(topic_id) {
+    // Validate topic id length.
+    if (topic_id.length > 64) return {success:false, invalid_input: true};
     // console.log("Delete topic:", topic_id);
-
     let { error } = await database.remove_topic(topic_id);
     if (error) {
         // console.log("Failed:", error); // Debug error.
@@ -72,6 +104,17 @@ async function DeleteTopic(topic_id) {
  */
 exports.UpdateTopic =
 async function UpdateTopic(userID, topic_id, topic_content) {
+    // Validate topic id length.
+    if (topic_id.length > 64) return {success:false, invalid_input: true};
+    // Sanitize topic content input.
+    topic_content = topic_content.replace(/[^a-zA-Z0-9\w\s]/gi, "");
+    if (!check_content(topic_content)) {
+        return {
+            success: false,
+            server_error: false,
+            invalid_input: true,
+        };
+    }
     // console.log("Update topic:", topic_id);
     let topic = {
         topic_id: topic_id,
@@ -79,8 +122,7 @@ async function UpdateTopic(userID, topic_id, topic_content) {
         user_posted: userID
     };
     let { error, update_time } = await database.update_topic(topic);
-    if (error) {
-        // Return error if query was unsucessfull.
+    if (error) {    // Return error if query was unsucessfull.
         // console.log("Failed:", error);
         return {
             success: false,
@@ -98,6 +140,7 @@ async function UpdateTopic(userID, topic_id, topic_content) {
 
 exports.GetTopic =
 async function GetTopic(topic_id) {
+    if (topic_id.length > 64) return {success: false, invalid_input: true};
     // console.log("Get topic:", topic_id);
     let {topic, error} = await database.get_topic(topic_id);
     if (error == database.errors.NO_RESPONSE) {
